busca por id com base na lista atualizada

+++ app/src/main/java/projetos/danilo/todolist/model/Task.kt
@@ -1,6 +1,7 @@
 package projetos.danilo.todolist.model
 
 import android.os.Parcelable
+import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import kotlinx.parcelize.Parcelize
@@ -10,7 +11,7 @@ import projetos.danilo.todolist.utils.Constants
 @Parcelize
 data class Task(
     @PrimaryKey(autoGenerate = true)
-    var id: Int,
+    var id: Long,
     var title: String,
     var description: String,
     var date: String,
